Share `usePopoverZIndex`
