+++ libs/ui/lib/listbox/Listbox.tsx
@@ -17,10 +17,13 @@ import { Listbox as Select } from '@headlessui/react'
 import cn from 'classnames'
 import type { ReactNode } from 'react'
 
-import { FieldLabel, SelectArrows6Icon, SpinnerLoader, TextInputHint } from '@oxide/ui'
-
-import { useIsInModal } from '../modal/Modal'
-import { useIsInSideModal } from '../side-modal/SideModal'
+import {
+  FieldLabel,
+  SelectArrows6Icon,
+  SpinnerLoader,
+  TextInputHint,
+  usePopoverZIndex,
+} from '@oxide/ui'
 
 export type ListboxItem<Value extends string = string> = {
   value: Value
@@ -83,13 +86,7 @@ export const Listbox = <Value extends string = string>({
   const selectedItem = selected && items.find((i) => i.value === selected)
   const noItems = !isLoading && items.length === 0
   const isDisabled = disabled || noItems
-  const isInModal = useIsInModal()
-  const isInSideModal = useIsInSideModal()
-  const zIndex = isInModal
-    ? 'z-modalDropdown'
-    : isInSideModal
-      ? 'z-sideModalDropdown'
-      : 'z-contentDropdown'
+  const zIndex = usePopoverZIndex()
 
   return (
     <div className={cn('relative', className)}>
